Added RowMappers class
